+++ numerapi/cli.py
@@ -82,6 +82,16 @@ def submission_ids(tournament=1):
 
 
 @cli.command()
+@click.option('--limit', default=20,
+              help='Number of items to return, defaults to 20')
+@click.option('--offset', default=0,
+            help='Number of items to skip, defaults to 0')
+def rankings(limit=20, offset=0):
+    """Get the overall rankings."""
+    click.echo(prettify(napi.get_rankings(limit=limit, offset=offset)))
+
+
+@cli.command()
 @click.option('--tournament', default=1,
               help='The ID of the tournament, defaults to 1')
 @click.option('--hours', default=24,