+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 # Changelog
 Notable changes to this project.
 
+## [dev]
+- added `get_rankings`
+
 ## [1.1.1] - 2018-06-06
 - added `get_tournaments`
 - added `tournament_name2number` and `tournament_number2name` to translate